refactor(i18n): migrate m10 consumers to public facade
