+++ src/content/video/sessionState.ts
@@ -1,4 +1,4 @@
-import type { PageI18nController } from '../../i18n';
+import type { PageI18nController } from '@i18n';
 import type { ReaderHighlightTheme } from '../../shared/types/options';
 import type { FragmentClipperOptions } from '../../shared/types/options';
 import type { VideoPlatform } from './utils';
