+++ src/options/components/controls/connectionTest.ts
@@ -5,7 +5,7 @@ import {
 import type { RestOptions } from '@shared/types/options';
 import type { VaultConfig } from '@shared/types/vault';
 import type { ConnectionTestResult } from '@shared/types/connection';
-import type { Messages } from '@i18n/locales';
+import type { Messages } from '@i18n';
 import { isAppError } from '@shared/errors';
 
 export interface ConnectionTesterConfig {
